+++ packages/vite/src/node/ssr/__tests__/ssrTransform.spec.ts
@@ -370,7 +370,7 @@ test('overwrite bindings', async () => {
           `const a = { inject }\n` +
           `const b = { test: inject }\n` +
           `function c() { const { test: inject } = { test: true }; console.log(inject) }\n` +
-          `const d = inject \n` +
+          `const d = inject\n` +
           `function f() {  console.log(inject) }\n` +
           `function e() { const { inject } = { inject: true } }\n` +
           `function g() { const f = () => { const inject = true }; console.log(inject) }\n`,
@@ -383,7 +383,7 @@ test('overwrite bindings', async () => {
     const a = { inject: __vite_ssr_import_0__.inject }
     const b = { test: __vite_ssr_import_0__.inject }
     function c() { const { test: inject } = { test: true }; console.log(inject) }
-    const d = __vite_ssr_import_0__.inject 
+    const d = __vite_ssr_import_0__.inject
     function f() {  console.log(__vite_ssr_import_0__.inject) }
     function e() { const { inject } = { inject: true } }
     function g() { const f = () => { const inject = true }; console.log(__vite_ssr_import_0__.inject) }
@@ -718,4 +718,21 @@ export default (function getRandom() {
   expect(
     (await ssrTransform(`export default (class A {});`, null, null)).code
   ).toMatchInlineSnapshot(`"__vite_ssr_exports__.default = (class A {});"`)
+})
+
+// #8002
+test('with hashbang', async () => {
+  expect(
+    (
+      await ssrTransform(
+        `#!/usr/bin/env node
+console.log("it can parse the hashbang")`,
+        null,
+        null
+      )
+    ).code
+  ).toMatchInlineSnapshot(`
+    "#!/usr/bin/env node
+    console.log(\\"it can parse the hashbang\\")"
+  `)
 })